+++ NEWS.md
@@ -3,6 +3,9 @@
 ## Minor Changes
 - updated the alias selection function within `input_processing()` so that an alias that is not already present is selected
 - updated the min-max scaling (controlled by `scale_vals`) in `color_kegg_pathway()`, the default is now `scale_vals=TRUE`
+- updated the `term_gene_heatmap()` function so that legend title is shown and can be customized
+- updated the `term_gene_heatmap()` function so that coloring is proper when no change values are provided in `genes_df`
+
 
 # pathfindR 1.6.3